requirements.txt
fastapi
uvicorn
gunicorn
sqlmodel
psycopg[binary]
requests
beautifulsoup4
python-dotenv
dateparser
pydantic
